// links-soma/app/src/features/bi/types/schema/edit-view-form.ts
import { z } from "zod";
import { result_views } from "../../../../db/schema";
import { parameterSchema } from "./parameter";

export const editViewFormSchema = z.object({
  dataSetResultId: z.coerce.number().optional(), // UI上はundefinedを許容
  title: z.string(),
  style: z.enum(result_views.style.enumValues),
  unit: z.enum(result_views.unit.enumValues),
  parameters: z.array(parameterSchema),
});